Enhancement: Enable play/pause via SIGUSR1
As noted in PR #3, Anipaper does not provide any arbitrary pause
mechanism, but only automatic pause when the screen is filled by
other windows and the wallpaper becomes little or no longer visible.

This PR adds a pause support for the SIGUSR1 signal: whenever a
signal of type SIGUSR1 is received, anipaper pauses or plays the
current video. This is especially useful if another process wants
to control Anipaper.

From now on a user can pause Anipaper like this:
  $ pkill -SIGUSR1 anipaper # pause
  $ pkill -SIGUSR1 anipaper # play

Many thanks to @complexlogic for initially proposing this feature
via D-Bus, which served as the basis for the implementation of this
via signals.